Require IO::Uncompress::Gunzip
Without this, you can’t examine pages or execute scripts that happen
to be gzipped, which is common enough that we should not requrie the
users to know ahead of time whether the pages they visit need this.

t/scripts.t was already failing on system without it installed, which
is correct, because a script like that *should* be able to run for
WWW::Scripter to be considered functional (IMO).

Also, tell decoded_content to die, so the problem becomes apparent
immediately if this module is accidentally installed without that
prerequisite present.